requirements.txt
# Core dependencies
yt-dlp>=2023.3.4
selenium>=4.1.0
webdriver-manager>=3.8.0
opencv-python>=4.7.0.72
Pillow>=9.4.0

# VLM and ML dependencies
torch>=2.0.0  # Add the specific CUDA version if needed, e.g., torch==2.0.0+cu118
torchvision>=0.15.0
transformers>=4.27.0
# CLIP - needs to be installed from GitHub
git+https://github.com/openai/CLIP.git

# OCR dependencies
pytesseract>=0.3.10

# For image and video processing
ffmpeg-python>=0.2.0
numpy>=1.24.2
scipy>=1.10.1
moviepy>=1.0.3

# CLI tools
typer[all]>=0.7.0  # [all] includes rich and shellingham
rich>=13.3.0
pydantic>=1.10.0  # For data validation and settings management

# Storage and cloud integration
boto3>=1.26.0  # For AWS S3
google-cloud-storage>=2.7.0  # For Google Cloud Storage

# Web and API tools
requests>=2.28.0
fastapi>=0.95.0  # For API if needed
uvicorn>=0.21.0  # ASGI server for FastAPI

# Concurrency
aiohttp>=3.8.4
asyncio>=3.4.3

# Development dependencies
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0

# Documentation
mkdocs>=1.4.2
mkdocs-material>=9.1.0
mkdocstrings>=0.20.0
mkdocstrings-python>=0.8.3

# Jupyter for examples
jupyter>=1.0.0
ipykernel>=6.22.0
matplotlib>=3.7.1  # For visualization in notebooks
seaborn>=0.12.2  # For advanced visualization